Move SyncingResult out of the processor
